CI: use Rust v1.53.0 to build project
The project depends on the crate `lscolors', which needs a recent
version of the Rust compiler.  Otherwise, the following error is
emitted:

   error[E0658]: use of unstable library feature 'str_strip': newly added

The str.strip_prefix() became stable at the v1.45.0 release, but since
then 2 more bug-fix releases have been released.

Other crates create the following errors:

   error[E0658]: `if` is not allowed in a `const fn`
   error[E0658]: `match` is not allowed in a `const fn`

These features were added to the language at the release v1.49.0 of the
compiler.

Finally, after using Rust v1.49.0 an error is triggered that the
program needs v1.53.0 or later to build it.